# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jatek Food Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags:
        - drivers
      summary: List drivers
      parameters:
        - name: isAvailable
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags:
        - drivers
      summary: Get a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    patch:
      operationId: updateDriver
      tags:
        - drivers
      summary: Update driver info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBody"
      responses:
        "200":
          description: Updated driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
      security:
        - bearerAuth: []
      description: JWT required. Driver owner or admin only.
  /drivers/{id}/complete-profile:
    post:
      operationId: completeDriverProfile
      tags:
        - drivers
      summary: Complete a driver profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteDriverProfileBody"
      responses:
        "200":
          description: Driver profile completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers/{id}/earnings:
    get:
      operationId: getDriverEarnings
      tags:
        - drivers
      summary: Get driver earnings summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverEarnings"
      security:
        - bearerAuth: []
      description: JWT required. Driver owner or admin only.
  /drivers/{id}/location:
    patch:
      operationId: updateDriverLocation
      tags:
        - drivers
      summary: Update driver live location
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverLocationUpdateBody"
      responses:
        "200":
          description: Driver location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverLocation"
        "400":
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getDriverLocation
      tags:
        - drivers
      summary: Get driver live location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverLocationPublic"
        "400":
          description: Invalid driver id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Driver not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type:
            - string
            - "null"
        vehicleType:
          type:
            - string
            - "null"
        isVerified:
          type: boolean
        isAvailable:
          type: boolean
        totalDeliveries:
          type: integer
        rating:
          type:
            - number
            - "null"
        createdAt:
          type: string
          format: date-time
        vehiclePlate:
          type:
            - string
            - "null"
        nationalId:
          type:
            - string
            - "null"
        licenseNumber:
          type:
            - string
            - "null"
        photoUrl:
          type:
            - string
            - "null"
        latitude:
          type:
            - number
            - "null"
        longitude:
          type:
            - number
            - "null"
        locationUpdatedAt:
          type:
            - string
            - "null"
          format: date-time
        profileCompletedAt:
          type:
            - string
            - "null"
          format: date-time
      required:
        - id
        - userId
        - name
        - isVerified
        - isAvailable
        - totalDeliveries
        - createdAt
...
    UpdateDriverBody:
      type: object
      properties:
        vehicleType:
          type: string
        isAvailable:
          type: boolean
    DriverEarnings:
      type: object
      properties:
        today:
          type: number
        thisWeek:
          type: number
        thisMonth:
          type: number
        totalDeliveries:
          type: integer
        completedToday:
          type: integer
      required:
        - today
        - thisWeek
        - thisMonth
        - totalDeliveries
        - completedToday
...
    CompleteDriverProfileBody:
      type: object
      required:
        - vehicleType
        - vehiclePlate
        - nationalId
      properties:
        vehicleType:
          type: string
        vehiclePlate:
          type: string
          minLength: 3
        nationalId:
          type: string
          minLength: 4
        licenseNumber:
          type:
            - string
            - "null"
        photoUrl:
          type:
            - string
            - "null"
    DriverLocationUpdateBody:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
        longitude:
          type: number
    DriverLocation:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type:
            - number
            - "null"
        longitude:
          type:
            - number
            - "null"
        locationUpdatedAt:
          type:
            - string
            - "null"
          format: date-time
    DriverLocationPublic:
      type: object
      properties:
        latitude:
          type:
            - number
            - "null"
        longitude:
          type:
            - number
            - "null"
        locationUpdatedAt:
          type:
            - string
            - "null"
          format: date-time
        isAvailable:
          type: boolean
        name:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT returned by /auth/login, /auth/verify-otp, or /backend/login. Some HTML document endpoints also accept the same token as a token query parameter.